refactor: Move component file to (components)
